Linked up days of treatment and added react-input-mask
