+++ package.json
@@ -20,6 +20,7 @@
     "react-circular-progressbar": "^1.0.0",
     "react-dom": "^16.5.1",
     "react-dropzone": "^7.0.1",
+    "react-input-mask": "^2.0.4",
     "react-scripts": "1.1.5",
     "reakit": "^0.15.1",
     "reakit-theme-default": "^0.1.12"
